move configuration to top level.
